BC-9190 - Move teams things to teams module
