+++ apps/server/src/modules/authorization-rules/rules/team.rule.spec.ts
@@ -3,13 +3,13 @@ import {
 	AuthorizationHelper,
 	AuthorizationInjectionService,
 } from '@modules/authorization';
+import { teamFactory } from '@modules/teams/testing';
 import { User } from '@modules/user/repo';
 import { userFactory } from '@modules/user/testing';
 import { Test, TestingModule } from '@nestjs/testing';
 import { Permission } from '@shared/domain/interface';
 import { setupEntities } from '@testing/database';
 import { roleFactory } from '@testing/factory/role.factory';
-import { teamFactory } from '@testing/factory/team.factory';
 import { TeamRule } from './team.rule';
 
 describe('TeamRule', () => {
