+++ apps/server/src/modules/collaborative-storage/collaborative-storage.module.ts
@@ -2,8 +2,8 @@ import { LoggerModule } from '@core/logger';
 import { CollaborativeStorageAdapterModule } from '@infra/collaborative-storage';
 import { AuthorizationModule } from '@modules/authorization';
 import { RoleModule } from '@modules/role';
+import { TeamsRepo } from '@modules/teams/repo';
 import { Module } from '@nestjs/common';
-import { TeamsRepo } from '@shared/repo/teams';
 import { CollaborativeStorageController } from './controller';
 import { TeamMapper, TeamPermissionsMapper } from './mapper';
 import { CollaborativeStorageService } from './services';
