+++ apps/server/src/modules/collaborative-storage/mapper/team.mapper.spec.ts
@@ -1,7 +1,7 @@
+import { TeamEntity } from '@modules/teams/repo';
+import { teamFactory } from '@modules/teams/testing';
 import { Test, TestingModule } from '@nestjs/testing';
-import { TeamEntity } from '@shared/domain/entity';
 import { setupEntities } from '@testing/database';
-import { teamFactory } from '@testing/factory/team.factory';
 import { TeamMapper } from './team.mapper';
 
 describe('TeamMapper', () => {
