+++ apps/server/src/modules/oauth-provider/domain/service/id-token.service.spec.ts
@@ -1,14 +1,14 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { PseudonymService } from '@modules/pseudonym/service';
+import { TeamsRepo } from '@modules/teams/repo';
+import { teamFactory } from '@modules/teams/testing';
 import { externalToolFactory } from '@modules/tool/external-tool/testing';
 import { User } from '@modules/user/repo';
 import { UserService } from '@modules/user/service/user.service';
 import { userDoFactory } from '@modules/user/testing';
 import { Test, TestingModule } from '@nestjs/testing';
-import { TeamsRepo } from '@shared/repo/teams';
 import { setupEntities } from '@testing/database';
 import { pseudonymFactory } from '@testing/factory/domainobject';
-import { teamFactory } from '@testing/factory/team.factory';
 import { IdTokenCreationLoggableException } from '../error';
 import { IdToken, OauthScope } from '../interface';
 import { IdTokenService } from './id-token.service';
