+++ apps/server/src/modules/oauth-provider/oauth-provider.module.ts
@@ -1,9 +1,9 @@
 import { LoggerModule } from '@core/logger';
 import { PseudonymModule } from '@modules/pseudonym';
+import { TeamsRepo } from '@modules/teams/repo';
 import { ToolModule } from '@modules/tool';
 import { UserModule } from '@modules/user';
 import { Module } from '@nestjs/common';
-import { TeamsRepo } from '@shared/repo/teams';
 import { IdTokenService } from './domain/service/id-token.service';
 import { OauthProviderLoginFlowService } from './domain/service/oauth-provider.login-flow.service';
 import { OauthProviderServiceModule } from './oauth-provider-service.module';
