+++ apps/server/src/modules/server/admin-api-server.entity.imports.ts
@@ -13,6 +13,7 @@ import { RocketChatUserEntity } from '@modules/rocketchat-user/entity';
 import { RoomMembershipEntity } from '@modules/room-membership';
 import { FederalStateEntity, SchoolEntity, SchoolYearEntity } from '@modules/school/repo';
 import { Task } from '@modules/task/repo';
+import { TeamEntity } from '@modules/teams/repo';
 import { ContextExternalToolEntity, LtiDeepLinkTokenEntity } from '@modules/tool/context-external-tool/repo';
 import { ExternalToolEntity } from '@modules/tool/external-tool/repo';
 import { SchoolExternalToolEntity } from '@modules/tool/school-external-tool/repo';
@@ -26,7 +27,6 @@ import {
 	SchoolNews,
 	StorageProviderEntity,
 	Submission,
-	TeamEntity,
 	TeamNews,
 } from '@shared/domain/entity';
