+++ apps/server/src/modules/teams/domain/index.ts
@@ -1,0 +1,1 @@
+export * from './service';
