+++ apps/server/src/modules/teams/repo/index.ts
@@ -1,0 +1,2 @@
+export { TeamEntity, TeamProperties, TeamUserEntity } from './team.entity';
+export { TeamsRepo } from './teams.repo';
