+++ apps/server/src/modules/teams/repo/team.repo.integration.spec.ts
@@ -1,14 +1,14 @@
 import { NotFoundError } from '@mikro-orm/core';
 import { EntityManager, ObjectId } from '@mikro-orm/mongodb';
+import { teamFactory } from '@modules/teams/testing';
 import { Test, TestingModule } from '@nestjs/testing';
-import { TeamEntity, TeamUserEntity } from '@shared/domain/entity';
 import { EntityId } from '@shared/domain/types';
-import { TeamsRepo } from '@shared/repo/teams';
 import { cleanupCollections } from '@testing/cleanup-collections';
 import { MongoMemoryDatabaseModule } from '@testing/database';
 import { roleFactory } from '@testing/factory/role.factory';
-import { teamFactory } from '@testing/factory/team.factory';
 import { teamUserFactory } from '@testing/factory/teamuser.factory';
+import { TeamEntity, TeamUserEntity } from './team.entity';
+import { TeamsRepo } from './teams.repo';
 
 describe('team repo', () => {
 	let module: TestingModule;
