+++ apps/server/src/modules/teams/repo/teams.repo.ts
@@ -1,8 +1,9 @@
 import { ObjectId } from '@mikro-orm/mongodb';
 import { Injectable } from '@nestjs/common';
-import { Role, TeamEntity, TeamUserEntity } from '@shared/domain/entity';
+import { Role } from '@shared/domain/entity';
 import { EntityId } from '@shared/domain/types';
-import { BaseRepo } from '../base.repo';
+import { BaseRepo } from '@shared/repo/base.repo';
+import { TeamEntity, TeamUserEntity } from './team.entity';
 
 @Injectable()
 export class TeamsRepo extends BaseRepo<TeamEntity> {
