+++ apps/server/src/modules/teams/teams.module.ts
@@ -1,9 +1,9 @@
 import { LoggerModule } from '@core/logger';
 import { AuthorizationModule } from '@modules/authorization';
+import { TeamsRepo } from '@modules/teams/repo';
 import { Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
-import { TeamsRepo } from '@shared/repo/teams';
-import { TeamAuthorisableService, TeamService } from './service';
+import { TeamAuthorisableService, TeamService } from './domain';
 
 @Module({
 	imports: [CqrsModule, LoggerModule, AuthorizationModule],
