+++ apps/server/src/modules/teams/testing/index.ts
@@ -1,0 +1,1 @@
+export * from './team.factory';
