+++ apps/server/src/modules/video-conference/service/video-conference.service.spec.ts
@@ -15,6 +15,8 @@ import { RoomService } from '@modules/room';
 import { RoomMembershipService } from '@modules/room-membership';
 import { roomMembershipFactory } from '@modules/room-membership/testing';
 import { roomFactory } from '@modules/room/testing';
+import { TeamsRepo } from '@modules/teams/repo';
+import { teamFactory } from '@modules/teams/testing';
 import { UserService } from '@modules/user';
 import { User } from '@modules/user/repo';
 import { userDoFactory, userFactory } from '@modules/user/testing';
@@ -25,11 +27,9 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { VideoConferenceDO } from '@shared/domain/domainobject';
 import { Permission, RoleName, VideoConferenceScope } from '@shared/domain/interface';
 import { EntityId } from '@shared/domain/types';
-import { TeamsRepo } from '@shared/repo/teams';
 import { VideoConferenceRepo } from '@shared/repo/videoconference';
 import { setupEntities } from '@testing/database';
 import { roleFactory } from '@testing/factory/role.factory';
-import { teamFactory } from '@testing/factory/team.factory';
 import { teamUserFactory } from '@testing/factory/teamuser.factory';
 import { videoConferenceDOFactory } from '@testing/factory/video-conference.do.factory';
 import { BBBRole } from '../bbb';
