+++ apps/server/src/modules/video-conference/service/video-conference.service.ts
@@ -6,15 +6,14 @@ import { CourseService } from '@modules/course';
 import { CourseEntity } from '@modules/course/repo';
 import { Room, RoomService } from '@modules/room';
 import { RoomMembershipService } from '@modules/room-membership';
+import { TeamEntity, TeamsRepo, TeamUserEntity } from '@modules/teams/repo';
 import { UserService } from '@modules/user';
 import { User } from '@modules/user/repo';
 import { BadRequestException, ForbiddenException, Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { RoleReference, VideoConferenceDO, VideoConferenceOptionsDO } from '@shared/domain/domainobject';
-import { TeamEntity, TeamUserEntity } from '@shared/domain/entity';
 import { Permission, RoleName, VideoConferenceScope } from '@shared/domain/interface';
 import { EntityId } from '@shared/domain/types';
-import { TeamsRepo } from '@shared/repo/teams';
 import { VideoConferenceRepo } from '@shared/repo/videoconference';
 import { BBBRole } from '../bbb';
 import { ErrorStatus } from '../error';
