+++ apps/server/src/modules/video-conference/uc/video-conference-deprecated.uc.spec.ts
@@ -6,6 +6,8 @@ import { AuthorizationService } from '@modules/authorization';
 import { CourseService } from '@modules/course';
 import { CourseEntity } from '@modules/course/repo';
 import { LegacySchoolService } from '@modules/legacy-school';
+import { TeamEntity, TeamsRepo } from '@modules/teams/repo';
+import { teamFactory } from '@modules/teams/testing';
 import { UserService } from '@modules/user';
 import { UserDo } from '@modules/user/domain';
 import { User } from '@modules/user/repo';
@@ -14,14 +16,12 @@ import { BadRequestException, ForbiddenException, InternalServerErrorException }
 import { Test, TestingModule } from '@nestjs/testing';
 import { AuthorizableObject } from '@shared/domain/domain-object';
 import { RoleReference, VideoConferenceDO } from '@shared/domain/domainobject';
-import { Role, TeamEntity } from '@shared/domain/entity';
+import { Role } from '@shared/domain/entity';
 import { Permission, RoleName, VideoConferenceScope } from '@shared/domain/interface';
 import { EntityId } from '@shared/domain/types';
-import { TeamsRepo } from '@shared/repo/teams';
 import { VideoConferenceRepo } from '@shared/repo/videoconference';
 import { setupEntities } from '@testing/database';
 import { roleFactory } from '@testing/factory/role.factory';
-import { teamFactory } from '@testing/factory/team.factory';
 import {
 	BBBBaseMeetingConfig,
 	BBBCreateConfigBuilder,
