+++ apps/server/src/modules/video-conference/video-conference.module.ts
@@ -8,10 +8,10 @@ import { LegacySchoolModule } from '@modules/legacy-school';
 import { RoleModule } from '@modules/role';
 import { RoomModule } from '@modules/room';
 import { RoomMembershipModule } from '@modules/room-membership';
+import { TeamsRepo } from '@modules/teams/repo';
 import { UserModule } from '@modules/user';
 import { HttpModule } from '@nestjs/axios';
 import { Module } from '@nestjs/common';
-import { TeamsRepo } from '@shared/repo/teams';
 import { VideoConferenceRepo } from '@shared/repo/videoconference/video-conference.repo';
 import { BBBService } from './bbb';
 import { VideoConferenceDeprecatedController } from './controller';
