+++ apps/server/src/shared/domain/entity/index.ts
@@ -5,5 +5,4 @@ export * from './news.entity';
 export * from './role.entity';
 export * from './storageprovider.entity';
 export * from './submission.entity';
-export * from './team.entity';
 export * from './video-conference.entity';
