+++ apps/server/src/shared/domain/entity/news.entity.ts
@@ -2,10 +2,10 @@ import { Entity, Enum, Index, ManyToOne, Property } from '@mikro-orm/core';
 import { CourseEntity } from '@modules/course/repo';
 import { SchoolEntity } from '@modules/school/repo';
 import type { User } from '@modules/user/repo';
+import type { TeamEntity } from '../../../modules/teams/repo/team.entity';
 import { EntityId } from '../types';
 import { NewsTarget, NewsTargetModel } from '../types/news.types';
 import { BaseEntityWithTimestamps } from './base.entity';
-import type { TeamEntity } from './team.entity';
 
 export interface NewsProperties {
 	title: string;
