+++ apps/server/src/shared/domain/types/news.types.ts
@@ -1,6 +1,6 @@
 import type { CourseEntity } from '@modules/course/repo';
 import { SchoolEntity } from '@modules/school/repo';
-import type { TeamEntity } from '../entity/team.entity';
+import type { TeamEntity } from '../../../modules/teams/repo/team.entity';
 import { EntityId } from './entity-id';
 
 export enum NewsTargetModel {
