--- apps/server/src/shared/repo/teams/index.ts
@@ -1,1 +1,0 @@
-export * from './teams.repo';
